requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
PyMuPDF==1.23.8
python-dotenv==1.0.1 